Add files via upload
Application Deployment Using CI/CD Tools

Implemented a git workflow architecture and used Jenkins jobs to deploy a container on specific ports for different branches, including the develop branch for testing the Dockerfile, the master branch for testing the final version of the Dockerfile, and the production server for the final release of the product. The process was automated using Jenkins, Git, and Docker.